Made TagScanner constructing synchronous
